layout delete block options panel continued
git-svn-id: http://svn.concrete5.org/svn/concrete5@1742 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/elements/block_area_layout_delete_opts.php
@@ -9,21 +9,34 @@ global $c;
 
 <div>
 
-	<div style="margin-bottom:16px;"><?=t("What would you like to do with this layout's blocks?") ?></div>
-	
 	<input id="ccm_delete_layout_cvalID" name="ccm_delete_layout_cvalID" type="hidden" value="1" />
-	
-	
-	
-	<input name="ccm_delete_layout_mode" type="radio" value="0" checked="checked" /> <?=t("Move them to the parent area below")?><br /><br />
-	
-	<input name="ccm_delete_layout_mode" type="radio" value="1" /> <?=t("Delete them")?><br /><br />
-	
-	
-	<div class="ccm-buttons">
-		<a href="#" class="ccm-button-left cancel" onclick="jQuery.fn.dialog.closeTop()"><span><em class="ccm-button-close"><?=t('Cancel')?></em></span></a>
+
+	<? if($_REQUEST['hasBlocks']){ ?>
+
+		<div style="margin-bottom:16px;"><?=t("What would you like to do with this layout's blocks?") ?></div> 
 		
-		<a href="javascript:void(0)" onclick="deleteLayoutObj.deleteLayout($('input[name=ccm_delete_layout_mode]:checked').val())" class="ccm-button-right accept"><span><?=t('Remove Layout') ?></span></a>
-	</div>	 
+		
+		<input name="ccm_delete_layout_mode" type="radio" value="0" checked="checked" /> <?=t("Move them to the parent area below")?><br /><br />
+		
+		<input name="ccm_delete_layout_mode" type="radio" value="1" /> <?=t("Delete them")?><br /><br />
+		
+		
+		<div class="ccm-buttons">
+			<a href="#" class="ccm-button-left cancel" onclick="jQuery.fn.dialog.closeTop()"><span><em class="ccm-button-close"><?=t('Cancel')?></em></span></a>
+			
+			<a href="javascript:void(0)" onclick="deleteLayoutObj.deleteLayout($('input[name=ccm_delete_layout_mode]:checked').val())" class="ccm-button-right accept"><span><?=t('Remove Layout') ?></span></a>
+		</div>	 
+	
+	<? }else{ ?>
+
+		<div style="margin:8px 0px 16px 0px;"><?=t("Are you sure you want to delete this layout section?") ?></div>
+		
+		<div class="ccm-buttons">
+			<a href="#" class="ccm-button-left cancel" onclick="jQuery.fn.dialog.closeTop()"><span><em class="ccm-button-close"><?=t('Cancel')?></em></span></a>
+			
+			<a href="javascript:void(0)" onclick="deleteLayoutObj.deleteLayout(1)" class="ccm-button-right accept"><span><?=t('Remove Layout') ?></span></a>
+		</div>	
+
+	<? } ?>
 
 </div>
